Struct re_dev_tools::build_search_index::meili::SearchClient

pub struct SearchClient {
    url: String,
    master_key: String,
    agent: Agent,
}

Fields
url: Stringmaster_key: Stringagent: Agent
Implementations

impl SearchClient

pub fn connect(url: &str, master_key: &str) -> Result<Self>

Connect to Meilisearch.

master_key can be obtained via the Meilisearch could console, or set via the --master-key option when running a local instance.

pub fn index(&self, index: &str, documents: &[Document]) -> Result<()>

Create an index from documents.

If an index with the same name already exists, it is deleted first.

pub fn query( &self, index: &str, q: &str, limit: Option<usize> ) -> Result<Vec<Document>>

Query a specific index in the database.

fn index_exists(&self, index: &str) -> Result<bool>

fn create_index(&self, index: &str) -> Result<()>

fn add_or_replace_documents( &self, index: &str, documents: &[Document] ) -> Result<()>

fn delete_index(&self, index: &str) -> Result<()>

fn wait_for_task(&self, task: Task) -> Result<()>

fn check_master_key(&self) -> Result<()>

fn get(&self, path: &str) -> Request

GET {self.url}{path}

fn post(&self, path: &str) -> Request

POST {self.url}{path}

fn delete(&self, path: &str) -> Request

DELETE {self.url}{path}

fn request(&self, method: Method, path: &str) -> Request

fn request_with_url(&self, method: Method, url: &Url) -> Request