re_dev_tools::build_web_viewer::lib
Function build

pub fn build(
    profile: Profile,
    debug_symbols: bool,
    target: Target,
    build_dir: &Utf8Path,
    no_default_features: bool,
    features: &String,
) -> Result<()>

Build re_viewer as Wasm, generate .js bindings for it, and place it all into the build_dir folder.

If debug_symbols is set, debug symbols are kept even in release builds, allowing for better callstacks on panics, as well as in-browser profiling of the wasm.