re_dev_tools
Module build_examples

This build script collects all examples which should be part of our example page, and either runs them to produce .rrd files, or builds a manifest file which serves as an index for the files.

It identifies runnable examples by checking if they have channel set in their README.md frontmatter. The available values are:

main for simple/fast examples built on each PR and the main branch

nightly for heavier examples built once per day

An example may also specify args to be run with via the frontmatter build_args string array.

Re-exports

pub use example::Example;

pub use example::ExamplesManifest;

pub use example::Language;

Modules

example 🔒

Example collection and parsing.

install 🔒

manifest 🔒

rrd 🔒

snippets 🔒

wait_for_output 🔒

Structs

Args

Build examples and their manifest.

Enums

Cmd 🔒

Functions

main